loader: move overlay-specific functionality out to a new overlay.go
